added TimeAt and basic date structure
We need to handle times other than 'now' so we added `CET.TimeAt` to
allow us to create any arbitrary time. We followed the method signature
for `time.Date`, but we added some extra validation to stop it from
doing unexpected things. For example we can't call this with the month
set to 13.

We also added a basic date type to allow us to safely represent dates.
One of the problems we've had is that go has no builtin date type and we
made some mistakes converting from times to dates with timezones.

Co-authored-by: John Cinnamond <john.cinnamond@form3.tech>